pyproject.toml
[project]
name = "lla-material-icons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.0.0",
    "cairosvg>=2.7.0",
]
